Fixing 2855 Inadvertent zlib dependency
Disables the b64_zlib_yaml format if zlib cannot be loaded.

I've added a --no-zlib to make it possible to test this on a single
machine, but it might also be useful if someone finds themselves failing
to connect to a server that doesn't have zlib installed.

FactHandler' format is still hard-coded to YAML rather than using
facts.class.default_format

Signed-off-by: Jesse Wolfe <jes5199@gmail.com>