Fixing #2958 inconsistent errors in spec/integration/bin/puppetmasterd.rb
I was getting inconsistent errors in this file, and it's really two
separate issues but they're both trivial.

1) The spec launches ./sbin/puppetmasterd, which may have trouble
finding the puppet libraries. It could even load the system-installed
puppet rather than the files in the repository -- I think this effect is
why the tests were passing on Markus's machine, for example.

2) There's a race condition in this test, and puppetmasterd has gotten
slower to start up (perhaps due to the #2877 fix?)

Signed-off-by: Jesse Wolfe <jes5199@gmail.com>