Fixing a bunch of Rails tests
This is a rebase of Luke's patch, with the addition of a special case
due to a change in how we pass SQLite parameters to ActiveRecord.

Signed-off-by: Jesse Wolfe <jes5199@gmail.com>

Luke's original commit (e9f88ebbce7163b1):

These tests were brittle and are less so now, but they're a
touch less thorough, too - we used to test that values weren't
provided (i.e., sqlite3 didn't provide a server name) but we now
only test the values provided.

I think these are sufficient (and far) better tests, though.

Signed-off-by: Luke Kanies <luke@reductivelabs.com>