Expression denotations should be cachable
The denotation of a pure expression should not change over time; thus it
should be possible to compute the result once and cache (memoize) it.

Just making the change causes a few test failures, because the tests were
explicitly depending on the erroneous behaviour; most of tests broken tests
were fixed.  In general the problem was that the older unit tests were
constructing an Expression and then repeatedly evaluating it under changing
conditions, expecting the results to change.

One test, which I do not understand, was not fixed:

    Failure:
    test_storeandcollect(TestScope)
        [./language/scope.rb:235:in `test_storeandcollect'
         ./language/scope.rb:234:in `each'
         ./language/scope.rb:234:in `test_storeandcollect'
         ./language/scope.rb:229:in `times'
         ./language/scope.rb:229:in `test_storeandcollect'
    Did not find puppet.
    <nil> is not true.