Centralize & normalize the AST#each system
The present system uses a "centralized" each method on the AST branch class
which should be (but isn't) the ancestor of all AST nodes that contain other
AST nodes, but it depends on them extending it correctly, which they don't
always do, etc., etc.

This patch moves the method up to the base class, gives it enough smarts that
it should never need to be overridden, and then removes the ad hoc extensions
from the various subclasses that at least tried to do the right thing.