Expression denotations should not depend on scope
In a declarative language, expressions combine values to produce results in a
deterministic fashion (3+4=7, regardless of the context).  Expression trees
should not need to be passed a scope to determine their denotation.

We aren't quite at that pure state yet, but as expression nodes know their scope
when they are created (and presently retain that information) we can at least
pretend that we're pure at the interface and, rather than passing in a scope on
all calls to evaluate, safeevaluate, evaluate_match, etc. we can omit it and
use the property in the rare cases where it's needed.  This should also make
the subsequent finding and removing of those cases much easier.

To make this a standalone commit we move the scope assignent to the Expression
node creation which will suffice for tests and be refactored away when we add
the incarnation phase.  To make refactoring it away easier, we mark all of the
occurances we add by thowing unneeded parins arround the scope.  This works for
all but one case (evaluate_code), and for that we add a stub of incarnation.