[#4182] show_diff was broken for streamed file contents
show_diff was written assuming that a file's contents would be loaded
into memory. That's no longer true, for perfomance reasons.
This patch streams the file to a temporary file to take the
diff.
As a consequence, it means that when show_diff is on, files may get
streamed twice.