Fix #2927 Symbolic file modes
File modes can now be specified as symbolic strings, in the same formats
that are supported by GNU chmod.
If a new file is created (or if a file is replaced with a directory, or
vice versa), all unspecified permissions default to 'denied'.
Otherwise, a file's permissions are considered in-sync as long as chmod
with the specified mode would not cause any permission bits to change.
This means it's now possible to have "don't care" permissions:
For example,
  file{"example" : mode => "a-x"}
will remove execute permissions, but leave it possible to manage the
read and write permissions outside of puppet.